requirements.txt
Flask==2.3.3
Flask-WTF==1.2.1
WTForms==3.0.1
requests==2.31.0
beautifulsoup4==4.12.2
openai==1.6.1
python-dotenv==1.0.0
celery==5.3.4
redis==5.0.1
gunicorn==21.2.0
httpx==0.27.2
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
alembic==1.13.1
